Split plugin files into modules
